Tests/LibWeb/Text/input/css/computed-values-stats.html

<!DOCTYPE html>
<script src="../include.js"></script>
<div id="target">hello</div>
<script>
    test(() => {
        let stats = internals.computedValuesStats();
        println(`liveComputedValues is a number: ${typeof stats.liveComputedValues === "number"}`);
        println(`totalComputedValuesCreated is a number: ${typeof stats.totalComputedValuesCreated === "number"}`);
        println(`liveComputedValues > 0: ${stats.liveComputedValues > 0}`);
        println(`totalComputedValuesCreated >= liveComputedValues: ${stats.totalComputedValuesCreated >= stats.liveComputedValues}`);

        let before = internals.computedValuesStats().totalComputedValuesCreated;
        document.getElementById("target").style.color = "red";
        internals.updateStyle();
        let after = internals.computedValuesStats().totalComputedValuesCreated;
        println(`restyle creates new ComputedValues: ${after > before}`);
    });
</script>
